Swap the name of the VM and the interpreter
